Outsource redundant code due to type conversion for sql.Scan():
sql.Scan() which is applied for record scanning only
expects intermediate interface slices. To store strings
(the values of each column/attribute) we need to associate
each element of the empty interface slice with the address
of the corresponding elemnt in a string pointer slice.

Until now there was redundant code in each endpoint function
that needed it which violates the DRY rule.

Therefore I implemented createSinkSliceForRecords().